rules segregation and unit tests
