+++ dv-rule-missingField-001/tests/MissingFieldRule.expect.json
@@ -1,0 +1,68 @@
+{
+  "name": "PingOne DaVinci Linter",
+  "pass": false,
+  "errorCount": 2,
+  "rulePackResults": [
+    {
+      "pass": false,
+      "errorCount": 2,
+      "lintResults": [
+        {
+          "flowId": "1e6254f9afffba533cf8e2f1efff06a3",
+          "flowName": "test",
+          "pass": false,
+          "errorCount": 2,
+          "errors": [
+            {
+              "code": "dv-er-validFields-001",
+              "message": "Missing site key configuration for reCAPTCHA",
+              "type": "error",
+              "recommendation": "Configure the required site key value to enable proper reCAPTCHA functionality.",
+              "flowId": "1e6254f9afffba533cf8e2f1efff06a3",
+              "nodeId": "6ux9r8ys3o"
+            },
+            {
+              "code": "dv-er-validFields-002",
+              "message": "Incorrect 'Additional Fields in the Response' in Send Success JSON Response capability",
+              "type": "error",
+              "recommendation": "Ensure valid key/value pairs are configured under 'Additional Fields in the Response'.",
+              "flowId": "1e6254f9afffba533cf8e2f1efff06a3",
+              "nodeId": "ru1iq6nna5"
+            }
+          ],
+          "rulesApplied": [
+            "dv-rule-validFields-001"
+          ],
+          "ruleResults": [
+            {
+              "ruleId": "dv-rule-validFields-001",
+              "ruleDescription": "",
+              "pass": false,
+              "errorCount": 2,
+              "errors": [
+                {
+                  "code": "dv-er-validFields-001",
+                  "message": "Missing site key configuration for reCAPTCHA",
+                  "type": "error",
+                  "recommendation": "Configure the required site key value to enable proper reCAPTCHA functionality.",
+                  "flowId": "1e6254f9afffba533cf8e2f1efff06a3",
+                  "nodeId": "6ux9r8ys3o"
+                },
+                {
+                  "code": "dv-er-validFields-002",
+                  "message": "Incorrect 'Additional Fields in the Response' in Send Success JSON Response capability",
+                  "type": "error",
+                  "recommendation": "Ensure valid key/value pairs are configured under 'Additional Fields in the Response'.",
+                  "flowId": "1e6254f9afffba533cf8e2f1efff06a3",
+                  "nodeId": "ru1iq6nna5"
+                }
+              ]
+            }
+          ],
+          "rulesIgnored": []
+        }
+      ],
+      "rulesIgnored": true
+    }
+  ]
+}
