+++ rules/dv-rule-node-001/tests/NodeRule.expect.json
@@ -1,26 +1,18 @@
 {
   "name": "PingOne DaVinci Linter",
   "pass": false,
-  "errorCount": 4,
+  "errorCount": 2,
   "rulePackResults": [
     {
       "pass": false,
-      "errorCount": 4,
+      "errorCount": 2,
       "lintResults": [
         {
           "flowId": "47360733f8216ad35114e9a787ca83e0",
           "flowName": "Starter Flow",
           "pass": false,
-          "errorCount": 4,
+          "errorCount": 2,
           "errors": [
-            {
-              "code": "dv-er-node-011",
-              "flowId": "47360733f8216ad35114e9a787ca83e0",
-              "message": "Missing 'Property Name' for Output Fields in Custom HTML Template capability",
-              "nodeId": "u5khyxw07p",
-              "recommendation": "Add a 'Property Name' to define the purpose of the output field. This property can then be referenced by successive nodes in the flow.",
-              "type": "error"
-            },
             {
               "flowId": "47360733f8216ad35114e9a787ca83e0",
               "code": "dv-bp-node-001",
@@ -36,14 +28,6 @@
               "nodeId": "13xqy80afe",
               "recommendation": "The (13xqy80afe Http) node is missing a description. For optimal clarity, add a meaningful description.",
               "type": "best-practice"
-            },
-            {
-              "code": "dv-er-node-008",
-              "flowId": "47360733f8216ad35114e9a787ca83e0",
-              "message": "Incorrect 'Additional Fields in the Response' in Send Error JSON Response capability",
-              "nodeId": "13xqy80afe",
-              "recommendation": "Ensure valid key/value pairs are configured under 'Additional Fields in the Response'.",
-              "type": "error"
             }
           ],
           "rulesApplied": [
@@ -54,16 +38,8 @@
               "ruleId": "dv-rule-node-001",
               "ruleDescription": "Ensure nodes have names/titles",
               "pass": false,
-              "errorCount": 4,
+              "errorCount": 2,
               "errors": [
-                {
-                  "code": "dv-er-node-011",
-                  "flowId": "47360733f8216ad35114e9a787ca83e0",
-                  "message": "Missing 'Property Name' for Output Fields in Custom HTML Template capability",
-                  "nodeId": "u5khyxw07p",
-                  "recommendation": "Add a 'Property Name' to define the purpose of the output field. This property can then be referenced by successive nodes in the flow.",
-                  "type": "error"
-                },
                 {
                   "flowId": "47360733f8216ad35114e9a787ca83e0",
                   "code": "dv-bp-node-001",
@@ -79,14 +55,6 @@
                   "nodeId": "13xqy80afe",
                   "recommendation": "The (13xqy80afe Http) node is missing a description. For optimal clarity, add a meaningful description.",
                   "type": "best-practice"
-                },
-                {
-                  "code": "dv-er-node-008",
-                  "flowId": "47360733f8216ad35114e9a787ca83e0",
-                  "message": "Incorrect 'Additional Fields in the Response' in Send Error JSON Response capability",
-                  "nodeId": "13xqy80afe",
-                  "recommendation": "Ensure valid key/value pairs are configured under 'Additional Fields in the Response'.",
-                  "type": "error"
                 }
               ]
             }
