+++ rules/dv-rule-subflow-001/tests/SubflowRule1.expect.json
@@ -1,17 +1,17 @@
 {
   "name": "PingOne DaVinci Linter",
   "pass": false,
-  "errorCount": 2,
+  "errorCount": 3,
   "rulePackResults": [
     {
       "pass": false,
-      "errorCount": 2,
+      "errorCount": 3,
       "lintResults": [
         {
           "flowId": "4822d6c2f9f61d1924909c50fa6959bc",
           "flowName": "Test - Main Flow",
           "pass": false,
-          "errorCount": 2,
+          "errorCount": 3,
           "errors": [
             {
               "flowId": "6cc59b44515ba46f388c87ef52a00ec9",
@@ -19,6 +19,13 @@
               "message": "Incorrect or missing subflow",
               "type": "error",
               "recommendation": "There is an incorrect or missing subflow in the (6cc59b44515ba46f388c87ef52a00ec9) flow. Ensure that the appropriate subflow is selected and that the names match within the configuration."
+            },
+            {
+              "code": "dv-er-subflow-004",
+              "flowId": "4822d6c2f9f61d1924909c50fa6959bc",
+              "message": "Invalid subflow configuration: PingOne flow selected as subflow",
+              "nodeId": "m7lhu8f6jg",
+              "recommendation": "PingOne flows cannot be referenced as a subflow. Select a standard DaVinci flow as the subflow instead."
             },
             {
               "code": "dv-er-subflow-005",
@@ -37,7 +44,7 @@
               "ruleId": "dv-rule-subflow-001",
               "ruleDescription": "Checks for subflow name mismatches",
               "pass": false,
-              "errorCount": 2,
+              "errorCount": 3,
               "errors": [
                 {
                   "flowId": "6cc59b44515ba46f388c87ef52a00ec9",
@@ -45,6 +52,13 @@
                   "message": "Incorrect or missing subflow",
                   "type": "error",
                   "recommendation": "There is an incorrect or missing subflow in the (6cc59b44515ba46f388c87ef52a00ec9) flow. Ensure that the appropriate subflow is selected and that the names match within the configuration."
+                },
+                {
+                  "code": "dv-er-subflow-004",
+                  "flowId": "4822d6c2f9f61d1924909c50fa6959bc",
+                  "message": "Invalid subflow configuration: PingOne flow selected as subflow",
+                  "nodeId": "m7lhu8f6jg",
+                  "recommendation": "PingOne flows cannot be referenced as a subflow. Select a standard DaVinci flow as the subflow instead."
                 },
                 {
                   "code": "dv-er-subflow-005",
